+++ app/data/pending.go
@@ -1,0 +1,66 @@
+package data
+
+import (
+	"sync"
+	"time"
+
+	"github.com/ethereum/go-ethereum/common"
+)
+
+// PendingPool - Currently present pending tx(s) i.e. which are ready to
+// be mined in next block
+type PendingPool struct {
+	Transactions map[common.Hash]*MemPoolTx
+	Lock         *sync.RWMutex
+}
+
+// Count - How many tx(s) currently present in pending pool
+func (p *PendingPool) Count() uint64 {
+
+	p.Lock.RLock()
+	defer p.Lock.RUnlock()
+
+	return uint64(len(p.Transactions))
+
+}
+
+// Add - Attempts to add new tx found in pending pool into
+// harmony mempool, so that further manipulation can be performed on it
+//
+// If it returns `true`, it denotes, it's success, otherwise it's failure
+// because this tx is already present in pending pool
+func (p *PendingPool) Add(tx *MemPoolTx) bool {
+
+	p.Lock.Lock()
+	defer p.Lock.Unlock()
+
+	if _, ok := p.Transactions[tx.Hash]; ok {
+		return false
+	}
+
+	// Marking we found this tx in mempool now
+	tx.DiscoveredAt = time.Now().UTC()
+
+	// Creating entry
+	p.Transactions[tx.Hash] = tx
+
+	return true
+
+}
+
+// Remove - Removes already existing tx from pending tx pool
+// denoting it has been mined i.e. confirmed
+func (p *PendingPool) Remove(txHash common.Hash) bool {
+
+	p.Lock.Lock()
+	defer p.Lock.Unlock()
+
+	if _, ok := p.Transactions[txHash]; !ok {
+		return false
+	}
+
+	delete(p.Transactions, txHash)
+
+	return true
+
+}
